Add support for type case
Type CASE statements is an Oberon-07 language feature that was
inadvertently missed earlier. This is an extension to the CASE statement
that works based on the type of a pointer (or VAR record parameter).

The form of this is
  CASE r OF
    R1: r.r1field := 10
  | R2: r.r2field := 11
  | R3: r.r3field := 12
  END

Where R1, R2, and R3 extend R. This works a lot like saying
  IF r IS R1 THEN r(R1).r1field := 10
  ELSIF r IS R2 THEN r(R2).r2field := 11
  ELSIF r IS R3 THEN r(R3).r3field := 12
  END

But has the advantage that the type guard is not really inserted,
because it isn't really necessary. So, any use of R within the case
statement body is appropriately cast to the type, since it's known to be
safe.

This was relatively straightforward in the RISC code generator, but
required a bit of bookkeeping for the C code generator. When r is used
within a case statement, C needs to keep track of the fact that there is
a cast, and needs to know what the name of that cast should be.