Improve how the compiler generates Type CASE
In the inttial support for typecase, it was neceaary to keep track, for
the benefit of the C code generator, if a variable needs a cast. This
involved adding a new pointer to each expression, indicating whether or
not that IdentRef was a safe cast.

This was wasteful, but of memory and of compiler complexity. Instead, in
the rare cases when an identifier is part of a type CASE statement, the
compiler can insert a TypeGuard. Thus, the code generator doesn't need
to do any additional work, so special case code for casting was replaced
with type guard.

Then, the type guard code was enhanced, by indicating that some type
guard expressions (the ones in a type case) don't need to do a runtime
type assertion test.

The end result is smaller code generator code, and more importantly, a
significant savings in memory. Adding the extra type to every expression
was expensive, and removing it reduces the heap usage of the compiler by
112748 bytes when compiling itself. This is a more than 6% reduction in
memory usage.