Implement a primitive pretty printer
This is the initial version of a very primitive formatter (pretty
printer) for Oberon. It bascially does a parse, and as it finds tokens,
emits them again. This puts the parser in charge of whitespace and line
breaks.

This version isn't ready for serious use yet, because, most notably, it
throws out comments (since the lexer on which it depends ignores
comments. However, it can be used to format simple pieces of code.