Factor out a lexer for the pretty printer
The pretty printer now has a fork of the lexer, which handles
indentation and formatting for the pretty printer. This is mostly due to
the need to not throw out comments, which will be added soon.

With this change, there is no need for pretty printer specific functions
in the main compiler's lexer.