Preserve comments in the code formatter
This uses a very simple heuristic to add comments back into the source
code output. Since each token is essentially parsed one at a time, as
soon as the scanner encounters a comment, it prints the comment. This is
not as robust as returning a comment token, but does have the advantage
that it isn't necessary to put comments into the string pool.

The pretty printer uses an extremely simple heuristic: if the comment is
the first thing on a line, then it is given its own line (and indented
to match the current indent level). Otherwise (the comment is inline,
and typically at the end of a line) it is printed inline. This is good
enough to do a reasonable job much of the time.