Remove immediate operations for real numbers
The emulated RISC CPU had immediate mode operations like RADDI to add an
immediate constant to a REAL number. These were removed because they're
not realistic (a real number literal wouldn't fit into the opcode of an
instruction anyway. Instead, the constant is loaded into a register
first, which was already working.