Remove unused compiler diagnostics
DumpDecl printed a version of the AST in a C comment, which was useful
in early stages of code generation. However, that output was only
partially correct, and has grown out of date.

This reduces the size of the compiler substantially, from over 5900 lines
of code to around 5300, so nearly 10% of the compiler was dedicated to
unused diagnostic information, and probably considerable strings as
well.