Make byte access more portable
The previous version of the compiler used C code to manually read bytes
out of words using LDW and STW. This code was informative, but somewhat
non-portable because it assumed that the ints making up memory were
little endian, an assumption needed by neither Oberon nor the
architecture.

Instead, this now maintains two pointers, one to memory represented as
ints, and the other as uint8_t. So the byte operations have been
delegate to the C compiler, but it should do the right thing regardless
of int format.