Support RETURN statements anywhere inside PROCEDURES
In Oberon-07, RETURN is special, and MUST be directly at the end of a
PROCEDURE body. With this optional language extension, a RETURN
statement can appear anywhere in a PROCEDURE body, e.g.

  PROCEDURE UseX(x: INTEGER): INTEGER;
  BEGIN
    IF x < 0 THEN RETURN -1 END;
    IF x > 10 THEN RETURN 1000 END;
    (* Code can guarantee x >= 0 and x <= 10 if it wants *)
    RETURN x + 1
  END;

Note that a RETURN is still required at the end of a procedure, even if it
can not possibly be reached, e.g.

  PROCEDURE UseX(x: INTEGER): INTEGER;
  BEGIN
    IF x < 0 THEN RETURN -1 END;
    IF x >= 0 THEN RETURN 1 END;
    (* Not reachable, but required for semantic compatability *)
    RETURN x + 1
  END;

RETURN is not allowed from a PROCEDURE that does not return a value.