Add an INTEGER test (Exponentiation with mod)
This code was inspired by Computerphile YouTube video about modulo
arithmetic and the square and multiply algorithm:

https://www.youtube.com/watch?v=cbGB__V8MNk

The point is that you can do (x^y) MOD m even if x^y could be an
extremely large number, far larger than can fit in a 32 bit integer. The
algorithm relies on the fact that multiplications MOD m will keep numbers
small, and that you can take the modulo whenever is necessary.

This scheme for exponentiation tracks the exponent from MSB to LSB,
squaring the number (mod m) and multiplying as you go. Because the
number squared easily fits within an INTEGER, and because the MOD
reduces the range, this allows you to compute even large exponents
efficiently (in terms of time and storage, using only plain integers.)

To perform the bit-by-bit from most significant to least significant
operation, the binary bits are first reversed, so that they can easily
be shifted out of the least significant bit. This is definitely not the
most efficient way to do this, but it works.