Dramatically reduce default RISC memory usage.
The 64K-word heap for the RISC runner has been reduced to 1K words. This
means that the memory footprint for the RISC virtual machine is only 4K
by default. This is ample for all of the regression tests, and can be
overridden when compiling something bigger (like the self-hosting
compiler.)

This was motivated by the desire to have multiple RISC virtual machines
simultaneously, and to prove that the generated code can do reasonable
things even with very little memory.