Simplify opcode generation
There is a tremendous amount of code shared between the assembly output
(human readable) and the code output (meant for sending as a C literal
to the RISC interpreter). This takes a step toward simplifying that, by
centralizing and simplifying the output of opcodes. Now, there is only
one method for taking an opcode and printing its name. Switching from a
cascading IF to a CASE statement leads to slighty larger generate code,
but considerably faster execution.

There is still a bunch of duplication that can be removed and code that
can be simplified, but this is a reasoable first step.