Update assembly output for CMP instructions
The code generator had a tiny hack because
  CMP R0, R1

really ends up being an instruction like
  SUB R0, R0, R1

But it is implied that the return value is discarded.  Previously, for
eadability, the compiler showed the two argument version. Now, it
displays
  CMP R0, R0, R1

which may be less readable but is closer to what is actually happening.

This also makes the code ever-so-smaller because Gen2 and Gen3 no longer
need an op parameter.