Change buffer INIT to INIT_ARRAY
The concept of a "buffer" is a non-standard extension to Oberon-07, and
INIT seemed too non-specific (one can imagine a generic initialization
mechanism, which this isn't.) INIT allocates a buffer. Probably this
should be called NEW, which is what Oberon-2 does.

This temporary solution changes INIT to INIT_ARRAY to avoid INIT being a
reserved word.