Revert shift based optimization
In the previous commit, multiply and divide were replaced by their shift
equivalents if the value is being multiplied or divided by a power of
two. While this is a nice optimization, it only works properly for
positive numbers. For dividing negative numbers, this causes truncation
toward negative infinity, which is not typical for programming
languages. Because of this, the maze generator (and probably the string
pool in the lexer) behave differently depending on implementation, which
is undesirable.

This keeps the previous test cases, but resets to the old behavior just
with multiply and divide.