Fork the M6502 emulator
If the tests have not been built or copied to the build directory, the
Apple1 emulator will not compile properly, because it depends on the
M6502 module. For now, simply copy the code to the examples, but perhaps
in the future there should be a mechanism to have a search path for
modules.