Remove the need for the C bootstrap compiler
Previously, the compiler first used the C bootstrap to build a compiler
written in C. This compiler written in C then built the oberon
transpiler, written in Oberon, which takes Oberon source and produces C.
Then it would use the Oberon compiler to compile itself.

Instead, the generated code, that is, the output of calling
  ../build/compile Compiler.ob
Which is the Oberon compiler transpiled to C, is checked into
oberon_bootstrap/. Thus, the only thing that the current Makefile builds
when running from scratch is a native compiler binary from this
bootstrap.

The old process can still be used with
  make -f Makefile-cbootstrap

This new scheme checks in generated code, which is not ideal, and also
needs to be manually regenerated if the Oberon to C transpiler changes,
but it is an improvement, and demonstrates the compiler's self-hosting
nature.

Note that there is a minor bug in compiler-test having to do with real
numbers, but otherwise everything seems to work.