Fix compiler-test for oberon_bootstrap
The source of the error was that the C compiler uses double to represent
a REAL, whereas the RISC compiler uses a float. Because of this, the lex
test that uses the numeric constant 128.3 has slightly different
representations, because this cannot be represented by a binary decimal
fraction without repeating digits.

The problem is avoided by using 128.25 which has a representation that
works.