Fix compiler-test a more appropriate way
Now that OBERON_REAL can be compiler-defined, the original real test can
be used. Instead, the compiler-test.sh script builds a custom version of
the compiler using only real numbers, so the output is compatibile with
a RISC interpreted version of the compiler.

This way, the triple tests still pass, but the C codegen uses the
preferred double representation by default.