Fix C code generation bug with typeinfos
When a pointer to a type is passed to a function, the typeinfo parameter
is never needed, because the typeinfo is derived from the pointer type.
This was generating the proper code when calling procedures, but not in
the procedure defintion. Now, the function signature generated for
pointers (or VAR pointers) does not include the typeid, because that was
not used.