Fix RISC code generation bug with real constants
When the RISC codegen generated code for expressions such as 2.0 +
Func(x), the generated code was incorrect. The reason was that an
optimization for integer constants doesn't work properly for real
numbers. For integer constants, it is typically possible to avoid
loading the constant into a register, and then use the value later
(e.g., ADDI R0, R0, #2).

This is not possible for real numbers, which need to be loaded into
registers first. As such, the system was generating code for the RHS of
an expression before generating the LHS. While this is technically
correct, many parts of the sytem assume that the result of a binary
expression will ultimately end up in the LHS.