Add a 10PRINT example
This is inspired by the classic Commodore 64 program, 10PRINT:

    10 PRINT CHR$(205.5+RND(1)); : GOTO 10

In Commodore BASIC, this alternates between the diagonal lines '/' and
'\' randomly. The end result looks like a maze printed on a diagonal.

This new test wasn't an attempt at code golfing (this program could
easily be much smaller) but was an excuse for me to experiment with a
Linear Feedback Shift Register (LFSR), tapping two of the bits (2 and
6), XORing them, shifting the word right, and adding the XORd bit on the
left. The end result should be a periodic number.

I picked 2 and 6 randomly, and, as it turns out, poorly, because the
output doesn't look great. However, it is a first start.