Improvie LFSR implementation
As discovered in the last 10PRINT experiment, choosing taps arbitrarily
for an XOR does not yield great results. Instead, the number and
position of the taps matters.

This new implementation was inspired by the Fibonacci XOR technique
described on wikipedia, using 11, 13, 14, and 16 (or, counting from LSB
to MSB, 0, 2, 3, and 5). The result ought to have a period of 65535 and
seems to produce better output.