Stop regenerating the compiler every time
Even if regen wasn't producing a different output, regeneration after
every compilation was causing make to always rebuild the compiler, which
was unnecessarily slow.

I should probably update the regen script to regenrate to the build
directory and only copy risc_code.txt if it has changed, but for now the
step is just disabled.