Swap the first and last set of random bits
The reason for this switch turns out to be interesting. I did a test of
generating three random boards in a row, and they turned out to be
exactly the same! This was a very odd coincidence, because generating
the random numbers with backtracking for a board exactly matched the
period of the random numbers. However, getting a different number first
throws off that period, so each new board (given those random numbers
and initial seed into the array) is different.