Remove experimental support for lower case identifiers
It was reliatively easy to add support for keywords like BEGIN to be in
lowercase like "begin", but, technically speaking, "begin" is a valid
Oberon-07 identifier.

The original goal was to make Oberon easier to read and to type, but in
retrospect, I think there is a better solution to this problem. Instead,
there should be editor support for auto upcasing identifiers (there was
once a vim plugin to do this using abbrev) or to use a code formatting
tool that could do the same thing.